+++ zed_nodelets/src/tools/include/sl_tools.h
@@ -1,9 +1,6 @@
-#ifndef SL_TOOLS_H
-#define SL_TOOLS_H
-
 ///////////////////////////////////////////////////////////////////////////
 //
-// Copyright (c) 2020, STEREOLABS.
+// Copyright (c) 2023, STEREOLABS.
 //
 // All rights reserved.
 //
@@ -21,6 +18,9 @@
 //
 ///////////////////////////////////////////////////////////////////////////
 
+#ifndef SL_TOOLS_H
+#define SL_TOOLS_H
+
 #include <ros/time.h>
 #include <sensor_msgs/Image.h>
 #include <sl/Camera.hpp>
@@ -29,18 +29,6 @@
 
 namespace sl_tools
 {
-/*! \brief Check if a ZED camera is ready
- * \param serial_number : the serial number of the camera to be checked
- */
-int checkCameraReady(unsigned int serial_number);
-
-/*! \brief Get ZED camera properties
- * \param serial_number : the serial number of the camera
- */
-sl::DeviceProperties getZEDFromSN(unsigned int serial_number);
-
-std::vector<float> convertRodrigues(sl::float3 r);
-
 /*! \brief Test if a file exist
  * \param name : the path to the file
  */