+++ zed_wrapper/records/record_depth.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-rosbag record /camera/rgb/camera_info /camera/depth_registered/image_raw/compressedDepth /camera/rgb/image_rect_color/compressed /tf