+++ zed_wrapper/records/record_stereo.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-rosbag record /camera/left/camera_info /camera/right/camera_info /camera/left/image_raw/compressed /camera/right/image_raw/compressed /tf